OCR/requirements.txt
Pillow>=9.0.0
opencv-python>=4.5.0
paddlepaddle>=2.4.0
paddleocr>=2.6.0
numpy>=1.19.0
requests>=2.25.0
python-dotenv>=0.19.0 